put notifications on the correct files
